refactor(Hand): Player, Dealer의 List<Card> 필드를 Hand 일급 컬렉션으로 변경
