+++ src/main/java/domain/Hand.java
@@ -1,0 +1,62 @@
+package domain;
+
+import java.util.ArrayList;
+import java.util.List;
+
+public class Hand {
+
+    private static final int INITIAL_HAND_SIZE = 2;
+    private static final int ACE_HIGH_LOW_DIFF = 10;
+    private static final int BUST_THRESHOLD = 21;
+
+    private final List<Card> cards = new ArrayList<>();
+
+    public void addInitialCards(List<Card> cards) {
+        cards.forEach(this::addCard);
+    }
+
+    public void addCard(Card card) {
+        cards.add(card);
+    }
+
+    public int calculateScore() {
+        int cardScore = calculateRawScore();
+        int aceCount = countAce();
+
+        for (int i = 0; i < aceCount; i++) {
+            cardScore = adjustForAce(cardScore);
+        }
+
+        return cardScore;
+    }
+
+    private int calculateRawScore() {
+        return cards.stream()
+                .mapToInt(Card::getScore)
+                .sum();
+    }
+
+    private int countAce() {
+        return (int) cards.stream().filter(Card::isAce).count();
+    }
+
+    private int adjustForAce(int cardScore) {
+        if (cardScore > BUST_THRESHOLD) {
+            return cardScore - ACE_HIGH_LOW_DIFF;
+        }
+        return cardScore;
+    }
+
+    public boolean isBust() {
+        return calculateScore() > BUST_THRESHOLD;
+    }
+
+    public boolean isBlackjack() {
+        return (calculateScore() == BUST_THRESHOLD) && (cards.size() == INITIAL_HAND_SIZE);
+    }
+
+    public List<Card> getCards() {
+        return List.copyOf(cards);
+    }
+
+}
